empezamos a divir auth.py, modificamos models.py para que solo tenga los modelos de usuario
